Update 07-exercise-configure-system-parameter-explore-metadata.md
Removed "restore" and "databaseName" parameters from the deployment command since they were not present in the deploy-postgresql-server.bicep file. Their removal did not affect the exercise but the "restore" parameter might have to be added to the biceps template like in the deploy.biceps file. Also removed links to images that are not in the repo anymore and fixed the download link for this exercise's script.